migrated core tests to a separate file
